PMM-8155 extract backup logic service to own pcakage
